# packages/server/openapi.snapshot.json
{
...
  "info": {
    "description": "Kryton server API",
    "title": "Kryton API",
    "version": "4.4.0-pre.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/shares/": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      },
                      "isFolder": {
                        "type": "boolean"
                      },
                      "ownerUserId": {
                        "type": "string"
                      },
                      "path": {
                        "type": "string"
                      },
                      "permission": {
                        "type": "string"
                      },
                      "sharedWithUserId": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "version": {
                        "maximum": 9007199254740991,
                        "minimum": -9007199254740991,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "id",
                      "ownerUserId",
                      "path",
                      "isFolder",
                      "sharedWithUserId",
                      "permission",
                      "createdAt",
                      "updatedAt",
                      "version"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List shares I own",
        "tags": [
          "collab"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "isFolder": {
                    "type": "boolean"
                  },
                  "path": {
                    "maxLength": 500,
                    "minLength": 1,
                    "type": "string"
                  },
                  "permission": {
                    "enum": [
                      "read",
                      "readwrite"
                    ],
                    "type": "string"
                  },
                  "sharedWithUserId": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "path",
                  "sharedWithUserId",
                  "permission"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "isFolder": {
                      "type": "boolean"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "permission": {
                      "type": "string"
                    },
                    "sharedWithUserId": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "version": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "ownerUserId",
                    "path",
                    "isFolder",
                    "sharedWithUserId",
                    "permission",
                    "createdAt",
                    "updatedAt",
                    "version"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create a share",
        "tags": [
          "collab"
        ]
      }
    },
...
    "/api/shares/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Revoke a share",
        "tags": [
          "collab"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "permission": {
                    "enum": [
                      "read",
                      "readwrite"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "permission"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "isFolder": {
                      "type": "boolean"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "permission": {
                      "type": "string"
                    },
                    "sharedWithUserId": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "version": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "ownerUserId",
                    "path",
                    "isFolder",
                    "sharedWithUserId",
                    "permission",
                    "createdAt",
                    "updatedAt",
                    "version"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Update share permission",
        "tags": [
          "collab"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
}